Hitscan into examples
